API/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
numpy==1.16.2
Pillow==5.4.1
tensorflow==1.14.0